+++ src/components/layout/app-sidebar.tsx
@@ -221,7 +221,7 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
   }
 
   return (
-    <Sidebar variant="inset" {...props}>
+    <Sidebar className={"sidebar"} variant="inset" {...props}>
       <SidebarHeader>
         <TooltipProvider delayDuration={0}>
           <Tooltip>